requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.8
MarkupSafe==2.1.3
packaging==24.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.1.3
